Lift the OpenAPI cycle-cutting out of the Google vendoring script
`cutCycles`, `referenced`, and the `Spec` shape lived inside
`google/specs/vendor.ts` while Google was the only vendored spec. None of
the three is about Google: they exist because `mcp-from-openapi` inlines
`$ref`s, which makes a generated schema cost what its reference graph
costs rather than what the document says. A cycle is then unbounded
recursion, not a self-reference.

A second vendored spec is arriving and needs `cutCycles` more urgently
than Google does. Copying seventy lines into a second script is how the
two come to disagree about what a cycle is, so this moves them to
`providers/shared/openapi.ts` and has the Google script import them
unchanged.

`makeOpaque` deliberately stays behind. It writes a sentence into the
schema it replaces and that sentence points at Google's own reference
documentation, which makes it a Google function that happens to look
generic.

Verified by re-running `bun run vendor:google`: every committed spec is
byte-identical, so this is a move rather than a rewrite.
